requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu

fastapi==0.115.0
uvicorn[standard]==0.30.0
pandas==2.2.2
scikit-learn==1.5.2
SQLAlchemy==2.0.30
PyMySQL==1.1.1
python-dotenv==1.2.1
sentence-transformers==2.7.0
umap-learn==0.5.6
hdbscan==0.8.37
strawberry-graphql==0.219.0
